Add Availabiliy component and functionality
